% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_distance}
\alias{color_distance}
\title{Color distance}
\usage{
color_distance(color_1, color_2)
}
\arguments{
\item{color_1}{color 1 in hex or rgb format}

\item{color_2}{color 2 in hex or rgb format}
}
\value{
a distance number
}
\description{
calculate distance between two color in hex or rgb format. rgb
maxColorValue = 255.
}
