% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{complex_html_text}
\alias{complex_html_text}
\title{Complex HTML text}
\usage{
complex_html_text()
}
\description{
Interactively combine columns by HTML styles and record workflow as
reproducible code.
}
\details{
When you're done, the code performing this operation will be emitted
at the cursor position.
}
