% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{convert_01}
\alias{convert_01}
\title{Convert character data to 0/1}
\usage{
convert_01(object)
}
\arguments{
\item{object}{data frame or list}
}
\description{
In data frame and list, convert character and numberic data to 0/1.
}
