% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{convert_01_to_connect}
\alias{convert_01_to_connect}
\title{Convert 0/1 data to connection pairs}
\usage{
convert_01_to_connect(object)
}
\arguments{
\item{object}{data frame with 0/1 data}
}
\value{
a data frame with source and target connection information
}
\description{
If two column has more than 1 shared element then they have connection. Convert
0/1 data to connection pairs in long shape table. The 0-connection pairs are removed.
}
