% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{convert_range_to_node}
\alias{convert_range_to_node}
\title{Convert range to node id}
\usage{
convert_range_to_node(df, tree)
}
\arguments{
\item{df}{data frame with any type of id}

\item{tree}{tree file path}
}
\description{
Convert the data frame with range id to node id by mrca method.
}
