% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_get_color}
\alias{correct_get_color}
\title{correct_get_color}
\usage{
correct_get_color(str)
}
\arguments{
\item{str}{taxa string}
}
\value{
a vector of colors
}
\description{
correct_get_color. (Version 0.0.0.9000)
}
