% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_type}
\alias{correct_type}
\title{Correct type parameter}
\usage{
correct_type(str)
}
\arguments{
\item{str}{A character string representing a type name to correct.}
}
\value{
A character string of the corrected type name.
}
\description{
Correct type parameter input by string similarity and synonym mapping.
}
\examples{
correct_type("line")
}
