% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_hub}
\alias{create_hub}
\title{Create itol.hub Object}
\usage{
create_hub(
  tree,
  field_tree = NULL,
  seq = NULL,
  abundance = NULL,
  taxonomy = NULL,
  node_data = NULL,
  tip_data = NULL
)
}
\arguments{
\item{tree}{tree file}

\item{field_tree}{todo}

\item{seq}{todo}

\item{abundance}{todo}

\item{taxonomy}{todo}

\item{node_data}{todo}

\item{tip_data}{todo}
}
\value{
Returns a itol.hub object
}
\description{
create a new object for itol.hub
}
\examples{
TREE <- system.file("extdata", "tree_of_itol_templates.tree", package = "itol.toolkit")
create_hub(tree = TREE)

}
\concept{object}
