% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{create_theme}
\alias{create_theme}
\title{Create itol.theme Object}
\usage{
create_theme(unit = NULL, file = NULL, tree = NULL, ...)
}
\arguments{
\item{unit}{unit object}

\item{file}{template file}

\item{tree}{tree file}

\item{...}{Further arguments to be passed to subsequent functions.}
}
\value{
Returns a itol.theme object
}
\description{
create a new object for itol.theme
}
\concept{object}
