% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{df_merge}
\alias{df_merge}
\title{Merge two data frame}
\usage{
df_merge(df1, df2, by = "id")
}
\arguments{
\item{df1}{initial data frame}

\item{df2}{sub data frame}

\item{by}{key column}
}
\description{
merge sub data frame into initial data frame
}
