% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{df_tree}
\alias{df_tree}
\title{Data frame tree}
\usage{
df_tree(df, main = "col", order = "none", sep = "_")
}
\arguments{
\item{df}{input data frame}

\item{main}{col(default) or row}

\item{order}{none or value}

\item{sep}{default '_'}
}
\description{
build a tree by row and column names of a data frame.
}
