% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fa_read}
\alias{fa_read}
\title{Read fasta file}
\usage{
fa_read(file)
}
\arguments{
\item{file}{input file in fasta format}
}
\value{
a data frame with sequence id and sequence
}
\description{
Read the fasta format sequences file into data.frame
}
