% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fa_write}
\alias{fa_write}
\title{Write fasta file}
\usage{
fa_write(object, file, id = "seq_name", seq = "sequence", append = FALSE)
}
\arguments{
\item{object}{data.frame format data}

\item{file}{input file in fasta format}

\item{id}{id col}

\item{seq}{seq col}

\item{append}{append at the end of an already existing file}
}
\description{
Write the fasta format sequences file from data.frame. (Version 0.0.0.9000)
}
