% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_get_name}
\alias{file_get_name}
\title{Get file name}
\usage{
file_get_name(str, with_ext = TRUE, keep_dir = FALSE)
}
\arguments{
\item{str}{str}

\item{with_ext}{with ext or not}

\item{keep_dir}{keep file dir or not}
}
\value{
a character specifying the file name
}
\description{
Get file name from string
}
