% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{file_to_unit}
\alias{file_to_unit}
\title{Create itol.unit Object from file}
\usage{
file_to_unit(file, tree, ...)
}
\arguments{
\item{file}{template file}

\item{tree}{tree file}

\item{...}{Further arguments to be passed to subsequent functions.}
}
\value{
Returns a itol.unit object
}
\description{
create a new object for itol.unit
}
\concept{object}
