% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color}
\alias{get_color}
\title{get_color}
\usage{
get_color(n, set = "table2itol")
}
\arguments{
\item{n}{level length of a vector}

\item{set}{a character specifying the palette set name. In dedault,
table2itol is setted. The following choices are possible: wsanderson.}
}
\value{
a vector of colors
}
\description{
get color, support max length 40
}
