% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{gradient_color}
\alias{gradient_color}
\title{Generate gradient colors}
\usage{
gradient_color(n, start, mid = NULL, end = "#FFFFFF")
}
\arguments{
\item{n}{the length of vector}

\item{start}{start color in hex format}

\item{mid}{mid color in hex format, default is null.}

\item{end}{end color in hex format, default is white.}
}
\value{
a vector of gradient colors
}
\description{
generate a vector of gradient colors by start, mid, and end
colors.
}
