% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head_line}
\alias{head_line}
\title{head line}
\usage{
head_line(function_name)
}
\arguments{
\item{function_name}{parent function name}
}
\value{
a character specifying the template type
}
\description{
Head line for templates
}
