% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{hub_to_unit}
\alias{hub_to_unit}
\title{Create itol.unit Object from object}
\usage{
hub_to_unit(object, theme, key)
}
\arguments{
\item{object}{itol.hub object}

\item{theme}{itol.theme object}

\item{key}{key id of dataset name}
}
\value{
Returns a itol.unit object
}
\description{
create a new object for itol.unit
}
\concept{object}
