% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inbuilt_themes}
\alias{inbuilt_themes}
\title{inbuilt themes}
\format{
\subsection{\code{inbuilt_themes}}{

A list with 23 template themes:
\describe{
\item{COLLAPSE}{Default theme of collapse template}
\item{PRUNE}{Default theme of prune template}
\item{SPACING}{Default theme of spacing template}
\item{TREE_COLORS}{Default theme of tree colors template}
\item{DATASET_STYLE}{Default theme of style template}
\item{LABELS}{Default theme of labels template}
\item{DATASET_TEXT}{Default theme of text template}
\item{DATASET_COLORSTRIP}{Default theme of colorstrip template}
\item{DATASET_BINARY}{Default theme of binary template}
\item{DATASET_GRADIENT}{Default theme of gradient template}
\item{DATASET_HEATMAP}{Default theme of heatmap template}
\item{DATASET_SYMBOL}{Default theme of symbol template}
\item{DATASET_EXTERNALSHAPE}{Default theme of externalshape template}
\item{DATASET_DOMAINS}{Default theme of domains template}
\item{DATASET_SIMPLEBAR}{Default theme of simple bar template}
\item{DATASET_MULTIBAR}{Default theme of multi bar template}
\item{DATASET_BOXPLOT}{Default theme of box plot template}
\item{DATASET_LINECHART}{Default theme of line chart template}
\item{DATASET_PIECHART}{Default theme of pie chart template}
\item{DATASET_ALIGNMENT}{Default theme of alignment template}
\item{DATASET_CONNECTION}{Default theme of connection template}
\item{DATASET_IMAGE}{Default theme of image template}
\item{POPUP_INFO}{Default theme of popup info template}
...
}
}
}
\usage{
inbuilt_themes
}
\description{
Default themes learned from iTOL offical template examples.
}
\keyword{datasets}
