% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{itol.hub-class}
\alias{itol.hub-class}
\alias{itol.hub}
\title{The itol.hub Class}
\description{
The itol.hub object is an intermediate storage container used internally throughout the
integration procedure to hold bits of data that are useful downstream.
}
\section{Slots}{

\describe{
\item{\code{tree}}{a list of meta data table, usually raw, full, and analyze}

\item{\code{seq}}{identity of the active assay}

\item{\code{abundance}}{abundance}

\item{\code{taxonomy}}{taxonomy}

\item{\code{meta.data}}{other meta.data}

\item{\code{theme}}{itol theme}
}}

\concept{objects}
