% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{itol.theme-class}
\alias{itol.theme-class}
\alias{itol.theme}
\title{The itol.theme Class}
\description{
The itol.theme object is an intermediate storage container used internally throughout the
integration procedure to hold bits of data that are useful downstream.
}
\section{Slots}{

\describe{
\item{\code{type}}{a list of meta data table, usually raw, full, and analyze}

\item{\code{sep}}{identity of the active assay}

\item{\code{profile}}{abundance}

\item{\code{field}}{taxonomy}

\item{\code{common_themes}}{other meta.data}

\item{\code{specific_themes}}{itol theme}
}}

\concept{objects}
