% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_data_from_file}
\alias{learn_data_from_file}
\title{Learn object data from file}
\usage{
learn_data_from_file(object, file)
}
\arguments{
\item{object}{itol.hub object}

\item{file}{template file}
}
\description{
Learn itol.hub object data from template file.
}
