% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_data_from_files}
\alias{learn_data_from_files}
\title{Learn object data from files}
\usage{
learn_data_from_files(object, files = NULL, dir = NULL, pattern = ".", ...)
}
\arguments{
\item{object}{itol.hub object}

\item{files}{template files path}

\item{dir}{files path}

\item{pattern}{file name pattern in regex}

\item{...}{Further arguments to be passed to subsequent functions.}
}
\description{
Learn itol.hub object data from template file.
}
