% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_data_from_unit}
\alias{learn_data_from_unit}
\title{Learn object data from unit}
\usage{
learn_data_from_unit(object, unit)
}
\arguments{
\item{object}{itol.hub object}

\item{unit}{itol.unit object}
}
\value{
a itol.hub object containing new data from itol.unit object
}
\description{
Learn itol.hub object data from unit object.
}
