% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_data_from_unit_list}
\alias{learn_data_from_unit_list}
\title{Learn object data from units}
\usage{
learn_data_from_unit_list(object, units)
}
\arguments{
\item{object}{itol.hub object}

\item{units}{itol.unit object list}
}
\value{
a itol.hub object with new data from a list of itol.unit objects
}
\description{
Learn itol.hub object data from list of unit object.
}
