% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_df}
\alias{learn_df}
\title{Learn from tree}
\usage{
learn_df(tree, node = FALSE, tip = TRUE)
}
\arguments{
\item{tree}{Newick tree file}

\item{node}{output node label}

\item{tip}{output tip label}
}
\description{
Learn initial data frame from Newick format tree leaves.
}
