% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_line}
\alias{learn_line}
\title{Learn paramter}
\usage{
learn_line(lines, param, sep)
}
\arguments{
\item{lines}{file lines}

\item{param}{a charactor string of paramter key name}

\item{sep}{separator}
}
\description{
learn paramter name and values based on the key name in the
front of line.
}
