% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_separator}
\alias{learn_separator}
\title{Learn separator}
\usage{
learn_separator(lines = NULL, file = NULL)
}
\arguments{
\item{lines}{string vector}

\item{file}{template file}
}
\description{
Learn 3 types of separators: tab, space, and comma.
}
