% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_alignment}
\alias{learn_theme_alignment}
\title{Learn alignment}
\usage{
learn_theme_alignment(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of alignment parameters containing
}
\description{
learn alignment paramters as list
}
