% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_bar}
\alias{learn_theme_bar}
\title{Learn bar}
\usage{
learn_theme_bar(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{a character specifying the separator.}
}
\value{
a list of bar parameters containing
}
\description{
learn bar paramters as list
}
