% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_basic_plot}
\alias{learn_theme_basic_plot}
\title{Learn basic plot}
\usage{
learn_theme_basic_plot(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of basic plot parameters containing
}
\description{
learn basic plot paramters as list
}
