% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_basic_theme}
\alias{learn_theme_basic_theme}
\title{Learn basic theme}
\usage{
learn_theme_basic_theme(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of basic theme parameters containing
}
\description{
learn basic theme paramters as list
}
