% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_binary}
\alias{learn_theme_binary}
\title{Learn binary}
\usage{
learn_theme_binary(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of binary chart parameters containing
}
\description{
learn binary paramters as list
}
