% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_border}
\alias{learn_theme_border}
\title{Learn border}
\usage{
learn_theme_border(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of border parameters containing
}
\description{
learn border paramters as list
}
