% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_domain}
\alias{learn_theme_domain}
\title{Learn domain}
\usage{
learn_theme_domain(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of domain parameters containing
}
\description{
learn domain paramters as list
}
