% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_heatmap}
\alias{learn_theme_heatmap}
\title{Learn heatmap}
\usage{
learn_theme_heatmap(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn heatmap paramters as list
}
