% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_image}
\alias{learn_theme_image}
\title{Learn image}
\usage{
learn_theme_image(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn connection paramters as list
}
