% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_linechart}
\alias{learn_theme_linechart}
\title{Learn linechart}
\usage{
learn_theme_linechart(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn linechart paramters as list
}
