% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_piechart}
\alias{learn_theme_piechart}
\title{Learn piechart}
\usage{
learn_theme_piechart(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of pie chart parameters containing
}
\description{
learn piechart paramters as list
}
