% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_specific_themes}
\alias{learn_theme_specific_themes}
\title{Learn specific themes}
\usage{
learn_theme_specific_themes(lines, sep, type)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}

\item{type}{template type}
}
\description{
learn specific theme paramters as list
}
