% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longest_continuous_match}
\alias{longest_continuous_match}
\title{Calculate longest continuous match between input and target strings}
\usage{
longest_continuous_match(input_str, target_str)
}
\arguments{
\item{input_str}{A character string to match against the target string.}

\item{target_str}{A character string to compare with the input string.}
}
\value{
An integer value representing the longest continuous match.
}
\description{
This function calculates the longest continuous matching substring between an input string and a target string.
}
