% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{merge_unit}
\alias{merge_unit}
\title{Merge units}
\usage{
merge_unit(obj1, obj2)
}
\arguments{
\item{obj1}{a itol.unit object specifying the first unit}

\item{obj2}{a itol.unit object specifying the second unit}
}
\value{
a itol.unit object with merged data
}
\description{
Merge two itol.unit with same type. The second unit data will be
added into the first one.
}
