% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{+,itol.hub,itol.unit-method}
\alias{+,itol.hub,itol.unit-method}
\title{\itemize{
\item method for S4 class itol.hub and itol.unit
}}
\usage{
\S4method{+}{itol.hub,itol.unit}(e1, e2)
}
\arguments{
\item{e1}{An object of class itol.hub}

\item{e2}{An object of class itol.unit}
}
\description{
\itemize{
\item method for S4 class itol.hub and itol.unit
}
}
