% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_tree_file}
\alias{search_tree_file}
\title{Search tree file}
\usage{
search_tree_file(
  dir = getwd(),
  n = "first",
  method = "mtime",
  max_size = 10240
)
}
\arguments{
\item{dir}{a path with tree file and other template files}

\item{n}{'first', 'last', 'all'}

\item{method}{sort by 'mtime', 'ctime', 'atime', 'character'}

\item{max_size}{limit file size to accelerate searching}
}
\value{
a vector of characters specifying the file name
}
\description{
Search Newick format tree file in dir
}
