% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,itol.hub-method}
\alias{show,itol.hub-method}
\title{show method for S4 class itol.hub}
\usage{
\S4method{show}{itol.hub}(object)
}
\arguments{
\item{object}{An object of class itol.hub}
}
\description{
show method for S4 class itol.hub
}
