% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{sort_color}
\alias{sort_color}
\title{Sort colors}
\usage{
sort_color(colors, root = NULL, rev = FALSE, plot = FALSE)
}
\arguments{
\item{colors}{a vector of colors in hex format}

\item{root}{sort from root color}

\item{rev}{sort order by similarity(default) or difference}

\item{plot}{preview color order}
}
\value{
a vector of sorted colors
}
\description{
sort colors by similarity
}
