% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_groups}
\alias{template_groups}
\title{template groups}
\format{
\subsection{\code{template_groups}}{

A data frame with template group clustering reslut:
\describe{
\item{template}{All the 23 template types of iTOL}
\item{group}{5 clustring gourps:
Tree structure: This group only controls the topology of tree branch
merging, filtering, and spacing. There are no style and rich annotation
data, even though most of the annotation data only include single-column id
information and do not contain any dataset base information, sample
information, or common and specific style information. It is a particularly
simple type of template.
Theme style: This does not change any topology or add any text information
but only changes the color scheme, line type and width, and font style and
size of existing information. This is an extremely comprehensive and
diverse type of annotation information.
Text: This group contains any templates with added text information. With
super flexible and convenient annotation methods, users can modify even a
single character's style in HTML. Users can also modify the text annotation
style of nodes and branches in batch based on matching conditions in
itol.hub objects, which require regular expression replacement and
precise data filtering. This high-frequency data processing is difficult to
achieve and retain the workflow in the EXCEL-based editor.
Basic plot: This group contains basic visualization methods. From a
functional point of view, this is the most feature-rich class of templates.
The similarity of the parameters within this part is very high.The
structured and uniform organization of these templates can greatly reduce
code redundancy and the user workload of data organizing. Moreover,
boxplot, which is not a regular enough data annotation template, can be
automatedly manipulated in R. The lack of template data structure makes
using frequency unbalanced among research. Hence, the frequency of using
these low-frequency templates can be increased.
Advanced plot: Compared with the basic visualization methods, these
visualization methods contain more comprehensive data types and often
require third-party tools for input data processing. But they are the most
extensible type of visualization methods for iTOL.
}
...
}
}
}
\usage{
template_groups
}
\description{
Templates were clustered into 5 groups by parameter similarity.
}
\keyword{datasets}
