% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{train_theme}
\alias{train_theme}
\title{Train inbuilt theme}
\usage{
train_theme(dir = getwd())
}
\arguments{
\item{dir}{the path of tree file and template files}
}
\value{
replace the global variable inbuilt_themes
}
\description{
The inbuilt theme is the template of all output file and unit.
Using this function can train the inbuilt theme object by custom files.
}
