% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unite_rows}
\alias{unite_rows}
\title{Paste rows}
\usage{
unite_rows(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
a data frame with pasted row by same id
}
\description{
Paste rows group by key column
}
