% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{vector_tree}
\alias{vector_tree}
\title{Vector tree}
\usage{
vector_tree(vector, branch_length = 1)
}
\arguments{
\item{vector}{input vector}

\item{branch_length}{branch length, default 1.}
}
\description{
build a tree by a vector of characters.
}
