% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_data}
\alias{learn_data}
\title{Learn data from template file}
\usage{
learn_data(df1 = NULL, file, tree = NULL, ...)
}
\arguments{
\item{df1}{initial data frame}

\item{file}{template file}

\item{tree}{tree file}

\item{...}{Further arguments to be passed to subsequent functions.}
}
\value{
a list with two data frame of node and tip annotation data
}
\description{
Learn data from template file into data frame
}
