% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_align}
\alias{learn_theme_align}
\title{Learn align}
\usage{
learn_theme_align(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn connection paramters as list
}
