% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_alignment}
\alias{learn_theme_alignment}
\title{Learn alignment}
\usage{
learn_theme_alignment(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn alignment paramters as list
}
