% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_basic_plot}
\alias{learn_theme_basic_plot}
\title{Learn basic plot}
\usage{
learn_theme_basic_plot(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn basic plot paramters as list
}
