% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_binary}
\alias{learn_theme_binary}
\title{Learn binary}
\usage{
learn_theme_binary(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn binary paramters as list
}
