% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_border}
\alias{learn_theme_border}
\title{Learn border}
\usage{
learn_theme_border(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn border paramters as list
}
