% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_common_themes}
\alias{learn_theme_common_themes}
\title{Learn common themes}
\usage{
learn_theme_common_themes(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn common theme paramters as list
}
