% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_domain}
\alias{learn_theme_domain}
\title{Learn domain}
\usage{
learn_theme_domain(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn domain paramters as list
}
