% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_externalshape}
\alias{learn_theme_externalshape}
\title{Learn externalshape}
\usage{
learn_theme_externalshape(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of external shape parameters containing
}
\description{
learn connection paramters as list
}
