% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_heatmap}
\alias{learn_theme_heatmap}
\title{Learn heatmap}
\usage{
learn_theme_heatmap(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of heatmap parameters containing
}
\description{
learn heatmap paramters as list
}
