% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_label}
\alias{learn_theme_label}
\title{Learn label}
\usage{
learn_theme_label(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn label paramters as list
}
