% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_linechart}
\alias{learn_theme_linechart}
\title{Learn linechart}
\usage{
learn_theme_linechart(lines, sep)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}
}
\value{
a list of line chart parameters containing
}
\description{
learn linechart paramters as list
}
