% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_specific_themes}
\alias{learn_theme_specific_themes}
\title{Learn specific themes}
\usage{
learn_theme_specific_themes(lines, sep, type)
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{sep}{a character specifying the separator.}

\item{type}{template type}
}
\value{
a list of specific theme parameters containing
}
\description{
learn specific theme paramters as list
}
