% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_theme_strip_label}
\alias{learn_theme_strip_label}
\title{Learn strip label}
\usage{
learn_theme_strip_label(lines, sep)
}
\arguments{
\item{lines}{file lines}

\item{sep}{separator}
}
\description{
learn strip label paramters as list
}
