% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{line_clean}
\alias{line_clean}
\title{Filter out comments and empty lines}
\usage{
line_clean(lines = NULL, file = NULL)
}
\arguments{
\item{lines}{string vector}

\item{file}{template file}
}
\description{
Remove the lines start with # or without any information.
}
