% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{line_split}
\alias{line_split}
\title{Split lines into two parts}
\usage{
line_split(lines, param = "data")
}
\arguments{
\item{lines}{a vector of character strings from template file.}

\item{param}{"theme" or "data" for the theme paramters or the data lines}
}
\value{
a vector of character strings containing data or theme information
}
\description{
Split lines based on the data block marker
}
