% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{+,itol.hub,itol.unit-method}
\alias{+,itol.hub,itol.unit-method}
\title{plus method
add method for S4 class itol.hub and itol.unit}
\usage{
\S4method{+}{itol.hub,itol.unit}(e1, e2)
}
\arguments{
\item{e1}{An object of class itol.hub}

\item{e2}{An object of class itol.unit}
}
\value{
a itol.hub object with new data from itol.unit object
}
\description{
plus method
add method for S4 class itol.hub and itol.unit
}
